Update README.md to align with license
The GNU GPL v3 which Prism-Menu is licensed, does not permit restrictions on who can use, modify, and distribute GPL v3 code if the program is publicly available. This PR rectifies a section of the readme that misleads the user as to the terms of the GPL v3, and requests that the user reads the license.